+++ src/components/Checkbox/Checkbox.js
@@ -1,0 +1,35 @@
+import React from 'react';
+import './Checkbox.css';
+
+const Checkbox = ({
+    q,
+    answers,
+    values,
+    bindField,
+    collectAnswer,
+    content,
+}) => {
+    return (
+        <div>
+            {answers &&
+                answers.map((option, idx) => (
+                    <div key={option}>
+                        <input
+                            type="checkbox"
+                            id={`${q.slug}-${option}`}
+                            name={q.slug}
+                            required={q.required}
+                            value={values[idx]}
+                            {...bindField(q.slug)}
+                            onChange={(e) =>
+                                collectAnswer(q.slug, e.target.value)
+                            }
+                        />
+                        <label for={q.slug}>{option}</label>
+                    </div>
+                ))}
+        </div>
+    );
+};
+
+export default Checkbox;
