+++ src/components/PhoneNumber/PhoneNumber.css
@@ -1,0 +1,21 @@
+.TextInput {
+    border: 2px solid var(--color-foggyBlue);
+    border-radius: 10px;
+    font-size: 1rem;
+    padding: 10px 12px;
+    width: 100%;
+    outline: none;
+}
+
+.TextInput[touched='true']:focus {
+    border: 2px solid #0072ff;
+}
+
+.TextInput[data-touched='true']:invalid {
+    border: 2px solid var(--color-secondary);
+}
+
+.RequiredError {
+    display: none;
+    color: var(--color-secondary);
+}
