+++ src/components/RadioWithFollowUp/RadioWithFollowUp.css
@@ -1,0 +1,54 @@
+.RadioGroup .Radio {
+    margin: 0.5em;
+    font-size: 1rem;
+    display: grid;
+    grid-template-columns: min-content auto;
+    gap: 0.5rem;
+}
+
+.RadioGroup .RadioInput {
+    display: flex;
+}
+
+.RadioGroup .RadioInput input {
+    opacity: 0;
+    width: 0;
+    height: 0;
+}
+.RadioGroup .RadioInput input:focus + .RadioControl {
+    box-shadow: 0 0px 6px var(--color-overcastBlue);
+}
+
+.RadioGroup .RadioInput .RadioButton:checked + .RadioControl {
+    background: radial-gradient(
+        var(--color-overcastBlue) 50%,
+        rgba(255, 0, 0, 0) 51%
+    );
+}
+
+.RadioGroup .RadioControl {
+    display: block;
+    width: 1rem;
+    height: 1rem;
+    border-radius: 50%;
+    border: 1px solid var(--color-foggyBlue);
+    transform: translateY(-0.05em);
+}
+
+.RadioGroup .RadioInput .RadioButton:checked + .RadioControl::before {
+    transform: scale(1);
+}
+
+.RadioGroup .RadioLabel {
+    line-height: 1;
+}
+
+.RadioButton[data-touched='true']:invalid:first-of-type ~ .RequiredError {
+    display: inline;
+    margin-left: 8px;
+}
+
+.RequiredError {
+    display: none;
+    color: var(--color-secondary);
+}
