+++ src/components/TextInput/TextInput.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import './TextInput.css';
+
+const TextInput = ({ q, bindField, collectAnswer, content }) => {
+    return (
+        <>
+            <input
+                type="text"
+                name={q.slug}
+                required={q.required}
+                className="TextInput"
+                {...bindField(q.slug)}
+                onChange={(e) => collectAnswer(q.slug, e.target.value)}
+            />
+            <div className="RequiredError">*{content.errorMessage}</div>
+        </>
+    );
+};
+
+export default TextInput;
